Add real data to the waffle chart on vote log page
